feeat: index.html에 마이페이지 구현
