+++ src/main/resources/static/css/base.css
@@ -474,7 +474,7 @@ body * {
         justify-content: center;
     }
 
-    .action-buttons .button{
+    .action-buttons .button {
         width: 90px;
         margin: 10px;
     }
@@ -862,7 +862,6 @@ body * {
 }
 
 
-
 #write-post {
     width: 800px;
     height: 960px;
@@ -873,7 +872,7 @@ body * {
     position: relative;
     font-family: 'Nanum Gothic', sans-serif;
 
-    .dropdown, .title-input{
+    .dropdown, .title-input {
         width: 100%;
         padding: 5px 0;
         margin-bottom: 20px;
@@ -1101,3 +1100,63 @@ body {
     margin: 5px 0;
     white-space: nowrap;
 }
+
+/* 페이지 전체 레이아웃 */
+body {
+    font-family: Arial, sans-serif;
+    margin: 0;
+    padding: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: darkseagreen;
+}
+
+.container {
+    display: grid;
+    grid-template-columns: 1fr 3fr 1fr; /* 좌측, 메인, 우측 */
+    gap: 20px; /* 섹션 간의 간격 */
+    width: 90%;
+    max-width: 1200px;
+    margin-top: 20px;
+}
+
+/* 공통 섹션 스타일 */
+.section {
+    border: 1px solid #333;
+    border-radius: 10px;
+    padding: 20px;
+    background-color: #f5f5f5;
+}
+
+/* 상단바는 그리드 전체를 차지하도록 설정 */
+.menu-bar {
+    grid-column: 1 / -1;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 10px 20px;
+    border: 1px solid #333;
+}
+
+/* 중앙 메인 콘텐츠 (내 정보, 내 목표, 참여 챌린지) */
+#my-info, #my-objectives, #participated-challenges {
+    grid-column: 2 / 3; /* 중앙 열에 배치 */
+}
+
+/* 오른쪽 사이드바 (향후 추가할 섹션) */
+#right-sidebar {
+    grid-column: 3 / 4;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+}
+
+/* 좌측 사이드바 (향후 추가할 섹션) */
+#left-sidebar {
+    grid-column: 1 / 2;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+}
+
