+++ src/main/resources/templates/fragment/participated-challenge.html
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
-    <link href="/static/css/reset.css" rel="stylesheet">
-    <link href="/static/css/base.css" rel="stylesheet">
+    <link href="/css/reset.css" rel="stylesheet">
+    <link href="/css/base.css" rel="stylesheet">
 </head>
 <body>
 <div th:fragment="participated-challenge">
