+++ src/main/resources/templates/index.html
@@ -1,26 +1,49 @@
 <!DOCTYPE html>
 <html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
-  <meta charset="UTF-8">
-  <title>타임리프 레이아웃</title>
-  <link rel="stylesheet" th:href="@{/css/base.css}">
+    <meta charset="UTF-8">
+    <title>MomentTree</title>
+    <link rel="stylesheet" th:href="@{/css/base.css}">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet">
+    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
 </head>
 <body>
-<div class="logo" th:src="@{/images/logo.png}"></div>
-<input type="text" class="search-bar" placeholder="검색">
 
-<div class="menu-bar">
-  <div th:each="menu, iterStat : ${subMenus}" style="display: inline-flex; align-items: center;">
-    <div class="menu-item">
-      <span th:text="${menu.title}"></span>
-      <!-- 서브메뉴 -->
-      <div class="submenu">
-        <span class="submenu-item" th:each="item : ${menu.items}" th:text="${item}"></span>
-      </div>
+<!-- 로고 및 검색 바 -->
+<div class="logo" th:src="@{/images/logo.png}"></div>
+<input class="search-bar" placeholder="검색" type="text">
+
+<!-- 상단 메뉴바 -->
+<div class="container">
+    <div class="menu-bar">
+        <div style="display: inline-flex; align-items: center;" th:each="menu, iterStat : ${subMenus}">
+            <div class="menu-item">
+                <span th:text="${menu.title}"></span>
+                <!-- 서브메뉴 -->
+                <div class="submenu">
+                    <span class="submenu-item" th:each="item : ${menu.items}" th:text="${item}"></span>
+                </div>
+            </div>
+            <!-- 마지막 메뉴 뒤에는 divider가 나타나지 않게 설정 -->
+            <div class="divider" th:if="${!iterStat.last}"></div>
+        </div>
     </div>
-    <!-- 마지막 메뉴 뒤에는 divider가 나타나지 않게 설정 -->
-    <div class="divider" th:if="${!iterStat.last}"></div>
-  </div>
 </div>
+
+<!-- 메인 콘텐츠 영역 -->
+<main>
+    <!-- 동적 콘텐츠 삽입 영역 -->
+    <section th:fragment="content">
+        <!-- 내 정보, 목표, 챌린지 섹션을 이곳에 불러옴 -->
+        <div th:replace="fragment/mypage-profile.html :: mypage-profile"></div>
+        <div th:replace="fragment/my-objective.html :: my-objective"></div>
+        <div th:replace="fragment/participated-challenge.html :: participated-challenge"></div>
+    </section>
+</main>
+
+<footer>
+
+</footer>
+
 </body>
 </html>
