--- src/main/resources/templates/mypage.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html th:replace="layout.html :: mypage" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title th:text="'My Page'"></title>
-    <link href="/static/css/reset.css" rel="stylesheet">
-    <link href="/static/css/base.css" rel="stylesheet">
-</head>
-<body>
-
-<section th:fragment="mypageContent">
-    <!-- 내 정보, 목표, 챌린지 섹션을 이곳에 불러옴 -->
-    <div th:replace="fragment/mypage-profile.html :: mypage-profile"></div>
-    <div th:replace="fragment/my-objective.html :: my-objective"></div>
-    <div th:replace="fragment/participated-challenge.html :: participated-challenge"></div>
-</section>
-
-
-</body>
-</html>
